+++ sdk/assemblyscript/src/assembly/models.ts
@@ -7,8 +7,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import { JSON } from "json-as";
 import * as utils from "./utils";
-import { Model, ModelFactory, ModelInfo } from "@hypermode/models-as";
+
+type ModelInvoker = (modelName: string, inputJson: string) => string | null;
 
 // @ts-expect-error: decorator
 @external("hypermode", "lookupModel")
@@ -42,6 +44,52 @@ class ModusModelFactory implements ModelFactory {
     }
 
     return instantiate<T>(info);
+  }
+}
+
+export class ModelInfo {
+  constructor(
+    public readonly name: string,
+    public readonly fullName: string = name,
+  ) {}
+}
+
+export interface ModelFactory {
+  getModel<T extends Model>(modelName: string): T;
+}
+
+export abstract class Model<TInput = unknown, TOutput = unknown> {
+  static invoker: ModelInvoker | null = null;
+  protected constructor(public info: ModelInfo) {}
+
+  debug: boolean = false;
+
+  /**
+   * Invokes the model with the given input.
+   * @param input The input object to pass to the model.
+   * @returns The output object from the model.
+   */
+  invoke(input: TInput): TOutput {
+    if (!Model.invoker) {
+      throw new Error("Model invoker is not set.");
+    }
+
+    const modelName = this.info.name;
+    const inputJson = JSON.stringify(input);
+    if (this.debug) {
+      console.debug(`Invoking ${modelName} model with input: ${inputJson}`);
+    }
+
+    const outputJson = Model.invoker(modelName, inputJson);
+    if (!outputJson) {
+      throw new Error(`Failed to invoke ${modelName} model.`);
+    }
+
+    if (this.debug) {
+      console.debug(`Received output: ${outputJson}`);
+    }
+
+    return JSON.parse<TOutput>(outputJson);
   }
 }